refactor: cleanup UI components
Use a better foldering structure for reusable components.
