--- components/Content.module.css
@@ -1,3 +1,0 @@
-.root {
-  flex: 1 1 auto;
-}
