+++ pages/about.tsx
@@ -2,8 +2,7 @@ import * as React from 'react'
 import { NextPage } from 'next'
 import Link from 'next/link'
 
-import { Page, PageHeader, PageBody } from '~/components/Layout'
-import Content from '~/components/Content'
+import { Content, Page, PageHeader, PageBody } from '~/components/layouts'
 import relativeTime from '~/utils/relativeTime'
 
 const lastUpdated = '2018-09-25T19:30:01+07:00'
